refactor: base to static file, configurations and generation warning
