+++ config/jekyll/favicon.yml
