+++ config/jekyll/favicon/asset/mutable.yml
