+++ config/jekyll/favicon/asset/referenceable.yml
