+++ config/jekyll/favicon/asset/sourceable.yml
