+++ config/jekyll/favicon/static_file.yml
