+++ lib/jekyll-favicon.rb
@@ -5,6 +5,7 @@ require 'jekyll/favicon'
 require 'jekyll/favicon/configuration'
 require 'jekyll/favicon/version'
 require 'jekyll/favicon/utils'
+require 'jekyll/favicon/static_file'
 require 'jekyll/favicon/generator'
 require 'jekyll/favicon/tag'
 require 'jekyll/favicon/hooks'
